+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 18/07/2023
+
 ### Added
 
 - added switch `--no-pdf-unc` to `plot` subcommand to skip the time-consuming
@@ -420,7 +422,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - first release
 
-[Unreleased]: https://github.com/NNPDF/pineappl/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/NNPDF/pineappl/compare/v0.6.1...HEAD
+[0.6.1]: https://github.com/NNPDF/pineappl/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/NNPDF/pineappl/compare/v0.5.9...v0.6.0
 [0.5.9]: https://github.com/NNPDF/pineappl/compare/v0.5.8...v0.5.9
 [0.5.8]: https://github.com/NNPDF/pineappl/compare/v0.5.7...v0.5.8